04_Nutrition_Performance_Optimization/backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests==2.31.0
pytesseract==0.3.10
Pillow==10.1.0
tensorflow==2.15.0
pandas==2.1.3
numpy==1.26.2

